Friend page

+++ lib/ui/screens/my_friends/my_friends_page.dart
@@ -1,6 +1,11 @@
 import 'package:bonobo/services/auth.dart';
 import 'package:bonobo/services/database.dart';
+import 'package:bonobo/services/storage.dart';
+import 'package:bonobo/services/storage_path.dart';
 import 'package:bonobo/ui/common_widgets/list_item_builder.dart';
+import 'package:bonobo/ui/common_widgets/platform_alert_dialog.dart';
+import 'package:bonobo/ui/common_widgets/platform_exception_alert_dialog.dart';
+import 'package:bonobo/ui/models/gender.dart';
 import 'package:bonobo/ui/screens/friend/friend_page.dart';
 import 'package:bonobo/ui/screens/my_friends/friend_list_tile.dart';
 import 'package:bonobo/ui/screens/my_friends/set_friend_form.dart';
@@ -20,13 +25,33 @@ class MyFriendsPage extends StatelessWidget {
   final FirestoreDatabase database;
   final List<SpecialEvent> allSpecialEvents;
 
-  void _deleteFriend(Friend friend) {
-    database.deleteFriend(friend);
+  void _deleteFriend(BuildContext context, Friend friend) async {
+    final res = await PlatformAlertDialog(
+      title: "Delete?",
+      content: "Are you sure want to delete ${friend.name}?",
+      defaultAtionText: "Yes",
+      cancelActionText: "Cancel",
+    ).show(context);
 
-    for (SpecialEvent event in allSpecialEvents) {
-      if (event.friendId == friend.id) {
-        database.deleteSpecialEvent(event);
+    try {
+      if (res) {
+        final storageService =
+            FirebaseStorageService(uid: database.uid, friend: friend);
+
+        await storageService.deleteFriendProfileImage();
+        database.deleteFriend(friend);
+        for (SpecialEvent event in allSpecialEvents) {
+          if (event.friendId == friend.id) {
+            database.deleteSpecialEvent(event);
+          }
+        }
       }
+    } catch (e) {
+      await PlatformAlertDialog(
+        title: "Error",
+        content: "Something went wrong",
+        defaultAtionText: "Ok",
+      ).show(context);
     }
   }
 
@@ -85,6 +110,7 @@ class MyFriendsPage extends StatelessWidget {
       actionExtentRatio: 0.18,
       child: Container(
         child: FriendListTile(
+          backgroundImage: NetworkImage(friend.imageUrl),
           friend: friend,
           onTap: () => {
             FriendPage.show(
@@ -111,7 +137,7 @@ class MyFriendsPage extends StatelessWidget {
           caption: 'Delete',
           color: Colors.red,
           icon: Icons.delete,
-          onTap: () => _deleteFriend(friend),
+          onTap: () => _deleteFriend(context, friend),
         ),
       ],
     );
